SSL: fix MLP head and remove L2 normalization (#145)
* draft projection head per Update the projection head (normalization and size). #139

* reorganize comments in example fit config

* configurable stem stride and projection dimensions

* update type hint and docstring for ContrastiveEncoder

* clarify embedding_dim

* use the forward method directly for projected

* normalize projections only when fitting
the projected features saved during prediction is now *not* normalized

* remove unused logger

* refactor training code into translation and representation modules

* extract image logging functions

* use AdamW instead of Adam for contrastive learning

* inline single-use argument

* fix normalization

* fix MLP layer order

* fix output dimensions

* remove L2 normalization before computing loss

* compute rank of features and projections

* documentation

---------
