+++ applications/contrastive_phenotyping/contrastive_cli/predict.yml
@@ -6,7 +6,7 @@ trainer:
   num_nodes: 1
   precision: 32-true
   callbacks:
-    - class_path: viscy.light.embedding_writer.EmbeddingWriter
+    - class_path: viscy.representation.embedding_writer.EmbeddingWriter
       init_args:
         output_path: "/hpc/projects/intracellular_dashboard/viral-sensor/infection_classification/models/contrastive_tune_augmentations/predict/test_prediction_code.zarr"
     # edit the following lines to specify logging path
