+++ examples/virtual_staining/img2img_translation/solution.py
@@ -128,8 +128,8 @@ from torch.utils.tensorboard import SummaryWriter  # for logging to tensorboard
 # HCSDataModule makes it easy to load data during training.
 from viscy.data.hcs import HCSDataModule
 # Trainer class and UNet.
-from viscy.light.engine import MixedLoss, VSUNet
-from viscy.light.trainer import VSTrainer
+from viscy.translation.engine import MixedLoss, VSUNet
+from viscy.translation.trainer import VSTrainer
 # training augmentations
 from viscy.transforms import (NormalizeSampled, RandAdjustContrastd,
                               RandAffined, RandGaussianNoised,
