+++ viscy/translation/predict_writer.py
